LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/compilation-error-1-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '%'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '%'
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '%'

FAIL Test that syntax errors lead to SyntaxError events on window, and that exceptions are remembered. assert_equals: expected 5 but got 3
